adds support for custom attributes in members (issue #26)
also 
- improves assembly comparison 
- re-enables tests for attribute named parameters
