+++ Cecilifier.Core.Tests/Tests/Integration/MembersTestCase.cs
@@ -18,5 +18,11 @@ namespace Cecilifier.Core.Tests.Integration
         {
             AssertResourceTest($@"Members/Events/{testName}");
         }
+
+        [Test]
+        public void AttributesOnMembers()
+        {
+            AssertResourceTest(@"Members/AttributesOnMembers");
+        }
     }
 }
